# docs/openapi.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "Cloudflare Backend Boilerplate API",
		"version": "0.1.0",
		"description": "API-only Cloudflare Worker (Hono) for mobile clients. Generated from src/routes/* and src/lib/schemas.ts — see scripts/generate-openapi.ts."
	},
	"servers": [
		{
			"url": "http://127.0.0.1:8787",
			"description": "Local dev"
		}
	],
	"components": {
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"description": "Better Auth JWT access token (15m). Also accepts an opaque session/API token on some routes — see docs/README."
			}
		}
	},
...
	"paths": {
...
		"/api/v1/uploads": {
			"post": {
				"tags": [
					"Uploads"
				],
				"summary": "Upload a file to R2 (multipart/form-data field `file`, or JSON base64)",
				"responses": {
					"200": {
						"description": "Stored object key, URL (if public), size and MIME type",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"data": {}
									}
								}
							}
						}
					},
					"400": {
						"description": "Structured error",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "object",
											"properties": {
												"code": {
													"type": "string"
												},
												"message": {
													"type": "string"
												},
												"details": {}
											},
											"required": [
												"code",
												"message"
											]
										}
									},
									"required": [
										"error"
									]
								}
							}
						}
					},
					"401": {
						"description": "Structured error",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "object",
											"properties": {
												"code": {
													"type": "string"
												},
												"message": {
													"type": "string"
												},
												"details": {}
											},
											"required": [
												"code",
												"message"
											]
										}
									},
									"required": [
										"error"
									]
								}
							}
						}
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"fileBase64": {
										"type": "string",
										"minLength": 1
									},
									"mimeType": {
										"type": "string",
										"minLength": 1
									},
									"filename": {
										"type": "string"
									}
								},
								"required": [
									"fileBase64",
									"mimeType"
								],
								"additionalProperties": false
							}
						}
					}
				}
			}
		},
		"/api/v1/uploads/{key}/url": {
			"get": {
				"tags": [
					"Uploads"
				],
				"summary": "Get a time-limited URL for a private upload",
				"responses": {
					"200": {
						"description": "Signed/public URL and expiry",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"data": {}
									}
								}
							}
						}
					},
					"400": {
						"description": "Structured error",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "object",
											"properties": {
												"code": {
													"type": "string"
												},
												"message": {
													"type": "string"
												},
												"details": {}
											},
											"required": [
												"code",
												"message"
											]
										}
									},
									"required": [
										"error"
									]
								}
							}
						}
					},
					"401": {
						"description": "Structured error",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "object",
											"properties": {
												"code": {
													"type": "string"
												},
												"message": {
													"type": "string"
												},
												"details": {}
											},
											"required": [
												"code",
												"message"
											]
										}
									},
									"required": [
										"error"
									]
								}
							}
						}
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"name": "key",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/uploads/{key}": {
			"delete": {
				"tags": [
					"Uploads"
				],
				"summary": "Delete an uploaded file",
				"responses": {
					"200": {
						"description": "File deleted",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"data": {}
									}
								}
							}
						}
					},
					"400": {
						"description": "Structured error",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "object",
											"properties": {
												"code": {
													"type": "string"
												},
												"message": {
													"type": "string"
												},
												"details": {}
											},
											"required": [
												"code",
												"message"
											]
										}
									},
									"required": [
										"error"
									]
								}
							}
						}
					},
					"401": {
						"description": "Structured error",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "object",
											"properties": {
												"code": {
													"type": "string"
												},
												"message": {
													"type": "string"
												},
												"details": {}
											},
											"required": [
												"code",
												"message"
											]
										}
									},
									"required": [
										"error"
									]
								}
							}
						}
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"name": "key",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	}
}